+++ src/System.Net.Http/src/System/Net/Http/HttpClientHandler.netcoreapp.cs
