+++ src/System.Net.NameResolution/src/System/Net/NameResolutionPal.Win32.cs
