+++ src/System.Threading.ThreadPool/src/ApiCompatBaseline.uapaot.txt
@@ -1,0 +1,1 @@
+MembersMustExist : Member 'System.Threading.ThreadPool.QueueUserWorkItem<TState>(System.Action<TState>, TState, System.Boolean)' does not exist in the implementation but it does exist in the contract.
